:truck: Move location_provider.dart
to providers folder
