+++ lib/views/GetPrayerTime.dart
@@ -8,7 +8,7 @@ import 'package:waktusolatmalaysia/models/jakim_esolat_model.dart';
 import '../providers/settingsProvider.dart';
 import '../CONSTANTS.dart';
 import '../locationUtil/locationDatabase.dart';
-import '../locationUtil/location_provider.dart';
+import '../providers/location_provider.dart';
 import '../notificationUtil/notification_scheduler.dart';
 import '../notificationUtil/prevent_update_notifs.dart';
 import '../utils/DateAndTime.dart';
